stable_diffusion/requirements.txt
mlx>=0.11
huggingface-hub
regex
numpy
tqdm
Pillow
